requirements.txt
arrow==1.2.1
matplotlib==3.5.0
mosaik==3.1.0
mosaik_api==3.0.2
mosaik-householdsim
numpy==1.20.3
pandas==1.2.4
protobuf==3.20.1
psutil==5.9.0
pytest==6.2.5
pytest_check==1.0.5
setuptools==57.0.0
simpy==3.0.13
termcolor==1.1.0
